Update/Delete appointment scheduling

+++ src/main/java/com/mballem/curso/security/web/controller/MedicoController.java
@@ -77,8 +77,15 @@ public class MedicoController {
 	
 	@GetMapping("/id/{idMed}/excluir/especializacao/{idEsp}")
 	public String excluirEspecialidadePorMedico(@PathVariable("idMed") Long idMed, @PathVariable("idEsp") Long idEsp, RedirectAttributes attr) {
-		service.excluirEspecialidadePorMedico(idMed, idEsp);
-		attr.addFlashAttribute("sucesso", "Especialidade removida com sucesso!");
+		
+		if ( service.existeEspecialidadeAgendada(idMed, idEsp)) {
+			attr.addFlashAttribute("falhar", "Tem consultas agendamdas, exclusãonegada.");
+			
+		} else {
+			service.excluirEspecialidadePorMedico(idMed, idEsp);
+			attr.addFlashAttribute("sucesso", "Especialidade removida com sucesso!");
+		}
+		
 		return "redirect:/medicos/dados";
 	}
